pyproject.toml
[project]
name = "stop-slop"
version = "0.1.0"
description = "Classifies slop websites using AI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "pandas",
  "requests",
  "beautifulsoup4",
  "undetected-chromedriver",
  "pyyaml",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools", "uv"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 88
target-version = "py39"
fix = true
unsafe-fixes = true

[tool.ruff.lint]  # <-- moved here from top-level [tool.ruff]
extend-select = ["I", "N", "UP", "B", "C4", "SIM", "RUF"]

[tool.ruff.format]
quote-style = "double"

[tool.black]
line-length = 88
target-version = ['py39']
skip-string-normalization = false

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
mypy_path = "src"

[dependency-groups]
dev = ["ruff", "black", "mypy", "pre-commit", "types-requests", "pandas-stubs", "types-requests", "types-PyYAML", "types-beautifulsoup4"]
